+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- tui: improve formatting of output by using click library
+- chore: update deps
+- core: add a timeout to code and dependecies checks
 
 ## [0.10] - 2026-05-20
 - chore: update deps